Add ConversationEntity to OLlama integration
